refactor: update Label component file name
