+++ src/modules/exercise/presentation/ui/components/fields/ExerciseNameField.tsx
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 import { InputText } from '@/presentation/modules/form/components/fields/InputText';
-import { Label } from '@/presentation/modules/form/components/fields/LabelGroup';
+import { Label } from '@/presentation/modules/form/components/fields/Label';
 import type { ExerciseDTO } from '@/modules/exercise/application/dtos/exercise.dto';
 import type { ExerciseFormProps } from '@/modules/exercise/presentation/ui/schemas/exercise.schema';
