+++ src/modules/exercise/presentation/ui/components/fields/ExerciseWeightField.tsx
@@ -1,7 +1,7 @@
 import { useFormContext } from 'react-hook-form';
 import { InputNumber } from '@/presentation/modules/form/components/fields/InputNumber';
 import { inputNumberConfig } from '@/presentation/modules/form/config/input-config';
-import { Label } from '@/presentation/modules/form/components/fields/LabelGroup';
+import { Label } from '@/presentation/modules/form/components/fields/Label';
 import { ExerciseDTO } from '@/modules/exercise/application/dtos/exercise.dto';
 import type { ExerciseFormProps } from '@/modules/exercise/presentation/ui/schemas/exercise.schema';
