+++ src/modules/routine/presentation/ui/components/fields/RoutineDayField.tsx
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 import { InputNumber } from '@/presentation/modules/form/components/fields/InputNumber';
-import { Label } from '@/presentation/modules/form/components/fields/LabelGroup';
+import { Label } from '@/presentation/modules/form/components/fields/Label';
 import { inputNumberConfig } from '@/presentation/modules/form/config/input-config';
 import type { RoutineForm } from '@/modules/routine/presentation/ui/config/routine.schema';
