+++ src/modules/routine/presentation/ui/components/fields/RoutineNameField.tsx
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 import { InputText } from '@/presentation/modules/form/components/fields/InputText';
-import { Label } from '@/presentation/modules/form/components/fields/LabelGroup';
+import { Label } from '@/presentation/modules/form/components/fields/Label';
 import type { RoutineDTO } from '@/modules/routine/application/dtos/routine.dto';
 
 type Props = {
